+++ src/index.js
@@ -58,9 +58,6 @@ const nameSelector = (segs) => {
   return literals.join('-');
 };
 
-/**
- * Routing table.
- */
 export const router = new Router(nameSelector)
   .add('/auth/*', auth)
   .add('/discover', discover)
@@ -106,7 +103,7 @@ async function run(request, context) {
   if (!handler) {
     return new Response('', { status: 404 });
   }
-  const info = RequestInfo.create(request, variables);
+  const info = RequestInfo.create(request, router, variables);
   if (info.method === 'OPTIONS') {
     return new Response('', {
       status: 204,